+++ arcade/future/__init__.py
@@ -1,9 +1,8 @@
 from . import video
 from . import light
-from . import input
 from . import background
 from . import splash
 
 from .texture_render_target import RenderTargetTexture
 
-__all__ = ["video", "light", "input", "background", "RenderTargetTexture", "splash"]
+__all__ = ["video", "light", "background", "RenderTargetTexture", "splash"]